refactor: move UserRepository interface to user directory
